Fix: Some bugs related to polygon kernels and console output for the GPU CLI version

+++ src/ParameterParser.cpp
@@ -200,7 +200,7 @@ void ParameterParser::ChangeSimuParams(WorkerParams& params){
                 params.halfLife = std::get<1>(par);
                 break;
             default:
-                Log::console_error("Unknown SimuParam {}\n", std::get<0>(par));
+                Log::console_error("Unknown SimuParam {}\n", (size_t) std::get<0>(par));
         }
     }
 }
@@ -238,7 +238,7 @@ int ParameterParser::ChangeFacetParams(std::vector<std::shared_ptr<SimulationFac
                     facet.sh.temperature = std::get<2>(par);
                     break;
                 default:
-                    Log::console_error("Unknown FacetParam {}\n", std::get<1>(par));
+                    Log::console_error("Unknown FacetParam {}\n", (size_t)std::get<1>(par));
             }
         }
         else{
